Add files via upload
How to run the NBC.

Make sure to have the csv formatted so the categories are stored as string in the second column, and the text to analyse in the third. 

You can change the num_cats variable to alter the number of categories that can be guessed. 

You can specify the rows for training and testing through the limit variables in the beginning. By default lines 21 and 199 split the rows into odds and evens, so no overlap can occur.